garden_tracker_app/backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
Werkzeug
Flask-JWT-Extended
Flask-CORS
fastapi
uvicorn[standard]
SQLAlchemy
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
alembic
pydantic-settings